requirements.txt
fastapi>=0.115,<1.0
uvicorn>=0.34,<1.0
google-genai>=1.0,<2.0
apscheduler>=3.10,<4.0
python-dotenv>=1.0,<2.0
